requirements.txt
Django>=4.2
djangorestframework
requests
pyserial
django-simple-history
django-nested-admin
